site/requirements.txt
ecdsa
pyOpenSSL